# specs/openapi.app.json
{
  "components": {
    "schemas": {
...
      "Campaign": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "campaignId": {
            "title": "Campaignid",
            "type": "string"
          },
          "channel": {
            "$ref": "#/components/schemas/Product"
          },
          "claimedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimedat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "counts": {
            "$ref": "#/components/schemas/CampaignCounts"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "fanoutCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fanoutcursor"
          },
          "footer": {
            "default": "",
            "title": "Footer",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/CampaignKind"
          },
          "lastProgressAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastprogressat"
          },
          "listId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listid"
          },
          "listIds": {
            "items": {
              "type": "string"
            },
            "title": "Listids",
            "type": "array"
          },
          "mediaKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediakey"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "optOutMode": {
            "$ref": "#/components/schemas/OptOutMode",
            "default": "REPLY_STOP"
          },
          "quoteMicro": {
            "title": "Quotemicro",
            "type": "integer"
          },
          "recipients": {
            "items": {
              "type": "string"
            },
            "title": "Recipients",
            "type": "array"
          },
          "reservedMicro": {
            "title": "Reservedmicro",
            "type": "integer"
          },
          "scheduledAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledat"
          },
          "senderId": {
            "title": "Senderid",
            "type": "string"
          },
          "settledMicro": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settledmicro"
          },
          "shortenUrls": {
            "title": "Shortenurls",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/CampaignStatus"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "body",
          "campaignId",
          "channel",
          "counts",
          "createdAt",
          "kind",
          "name",
          "quoteMicro",
          "recipients",
          "reservedMicro",
          "senderId",
          "shortenUrls",
          "status",
          "userId",
          "username"
        ],
        "title": "Campaign",
        "type": "object"
      },
      "CampaignCounts": {
        "properties": {
          "delivered": {
            "default": 0,
            "title": "Delivered",
            "type": "integer"
          },
          "queued": {
            "default": 0,
            "title": "Queued",
            "type": "integer"
          },
          "recipients": {
            "title": "Recipients",
            "type": "integer"
          },
          "refused": {
            "default": 0,
            "title": "Refused",
            "type": "integer"
          },
          "sent": {
            "default": 0,
            "title": "Sent",
            "type": "integer"
          },
          "undelivered": {
            "default": 0,
            "title": "Undelivered",
            "type": "integer"
          }
        },
        "required": [
          "recipients"
        ],
        "title": "CampaignCounts",
        "type": "object"
      },
      "CampaignDraft": {
        "properties": {
          "body": {
            "default": "",
            "title": "Body",
            "type": "string"
          },
          "footer": {
            "default": "",
            "title": "Footer",
            "type": "string"
          },
          "listId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listid"
          },
          "listIds": {
            "items": {
              "type": "string"
            },
            "title": "Listids",
            "type": "array"
          },
          "mediaKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediakey"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "optOutMode": {
            "$ref": "#/components/schemas/OptOutMode",
            "default": "REPLY_STOP"
          },
          "product": {
            "$ref": "#/components/schemas/Product",
            "default": "SMS"
          },
          "senderId": {
            "default": "",
            "title": "Senderid",
            "type": "string"
          },
          "shortenUrls": {
            "default": false,
            "title": "Shortenurls",
            "type": "boolean"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          }
        },
        "title": "CampaignDraft",
        "type": "object"
      },
      "CampaignKind": {
        "enum": [
          "LIST",
          "QUICK"
        ],
        "title": "CampaignKind",
        "type": "string"
      },
      "CampaignPage": {
        "properties": {
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Campaign"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "CampaignPage",
        "type": "object"
      },
      "CampaignQuote": {
        "properties": {
          "costMicro": {
            "title": "Costmicro",
            "type": "integer"
          },
          "parts": {
            "title": "Parts",
            "type": "integer"
          },
          "recipients": {
            "title": "Recipients",
            "type": "integer"
          },
          "senderDisplay": {
            "title": "Senderdisplay",
            "type": "string"
          }
        },
        "required": [
          "costMicro",
          "parts",
          "recipients",
          "senderDisplay"
        ],
        "title": "CampaignQuote",
        "type": "object"
      },
      "CampaignReport": {
        "properties": {
          "campaign": {
            "$ref": "#/components/schemas/Campaign"
          },
          "clicks": {
            "default": 0,
            "title": "Clicks",
            "type": "integer"
          }
        },
        "required": [
          "campaign"
        ],
        "title": "CampaignReport",
        "type": "object"
      },
      "CampaignStatus": {
        "enum": [
          "CANCELLED",
          "DRAFT",
          "FAILED",
          "SCHEDULED",
          "SENDING",
          "SENT"
        ],
        "title": "CampaignStatus",
        "type": "string"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "OptOutMode": {
        "enum": [
          "REPLY_STOP",
          "UNSUBSCRIBE_LINK"
        ],
        "title": "OptOutMode",
        "type": "string"
      },
...
      "Product": {
        "enum": [
          "MMS",
          "MMS_AS_LINK",
          "SMS"
        ],
        "title": "Product",
        "type": "string"
      },
...
      "ScheduleRequest": {
        "properties": {
          "now": {
            "default": false,
            "title": "Now",
            "type": "boolean"
          },
          "sendAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sendat"
          }
        },
        "title": "ScheduleRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "txtlocal",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/app/campaigns": {
      "get": {
        "operationId": "list_campaigns_api_app_campaigns_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Product"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "List Campaigns"
      },
      "post": {
        "operationId": "create_campaign_api_app_campaigns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Create Campaign"
      }
    },
    "/api/app/campaigns/{campaignId}": {
      "delete": {
        "operationId": "delete_campaign_api_app_campaigns__campaignId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "schema": {
              "title": "Campaignid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Delete Campaign"
      },
      "get": {
        "operationId": "get_campaign_api_app_campaigns__campaignId__get",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "schema": {
              "title": "Campaignid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Get Campaign"
      },
      "put": {
        "operationId": "save_campaign_api_app_campaigns__campaignId__put",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "schema": {
              "title": "Campaignid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Save Campaign"
      }
    },
    "/api/app/campaigns/{campaignId}/cancel": {
      "post": {
        "operationId": "cancel_campaign_api_app_campaigns__campaignId__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "schema": {
              "title": "Campaignid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Cancel Campaign"
      }
    },
    "/api/app/campaigns/{campaignId}/duplicate": {
      "post": {
        "operationId": "duplicate_campaign_api_app_campaigns__campaignId__duplicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "schema": {
              "title": "Campaignid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Duplicate Campaign"
      }
    },
    "/api/app/campaigns/{campaignId}/quote": {
      "post": {
        "operationId": "quote_campaign_api_app_campaigns__campaignId__quote_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "schema": {
              "title": "Campaignid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignQuote"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Quote Campaign"
      }
    },
    "/api/app/campaigns/{campaignId}/report": {
      "get": {
        "operationId": "report_campaign_api_app_campaigns__campaignId__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "schema": {
              "title": "Campaignid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Report Campaign"
      }
    },
    "/api/app/campaigns/{campaignId}/schedule": {
      "post": {
        "operationId": "schedule_campaign_api_app_campaigns__campaignId__schedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "schema": {
              "title": "Campaignid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Schedule Campaign"
      }
    },
...
    "/api/u/{token}": {
      "get": {
        "operationId": "unsubscribe_api_u__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Unsubscribe"
      }
    },
...
  }
}